pyproject.toml
[project]
name = "iplanrio-desafio-data-eng"
version = "0.1.0"
description = ""
readme = "README.md"
authors = [
    { name = "tricktx", email = "patrick.tx@hotmail.com.br" },
]
requires-python = ">=3.10.12"

dependencies = [
    "pandas>=2.2.2",
    "requests>=2.32.3",
    "bs4>=0.0.2",
    "numpy>=1.26.4",
    "dbt-core>=1.8.1",
    "tqdm==4.66.2",
    "google-auth==2.22.0",
    "google-auth-oauthlib>=1.0.0",
    "google-cloud-bigquery>=3.11.4",
    "google-cloud-bigquery-storage>=2.19.1",
    "google-cloud-storage>=2.9.0",
    "prefect>=3.4.25",
    "prefect-gcp>=0.6.11",
    "prefect-github>=0.3.2",
    "prefect-dbt>=0.7.0",
    "prefect-docker==0.7.0",
    "dbt-bigquery==1.9.1",

]

[project.optional-dependencies]
dev = [
    "ipykernel>=6.29.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]